Move xml-related functions into xml.{cpp,h}
